Complete user-managed browser web extensions
