+++ Sources/AppDelegate+BrowserWebExtensionShortcutRouting.swift
@@ -1,6 +1,64 @@
 import AppKit
 
+/// Whether the responder chain passes through a browser web view, so
+/// web-extension shortcut commands (e.g. Bitwarden autofill) may claim the event.
+func cmuxRespondersContainBrowserWebView(_ responder: NSResponder?) -> Bool {
+    var current: NSResponder? = responder
+    while let next = current {
+        if next is CmuxWebView { return true }
+        if let view = next as? NSView {
+            var ancestor: NSView? = view.superview
+            while let candidate = ancestor {
+                if candidate is CmuxWebView { return true }
+                ancestor = candidate.superview
+            }
+        }
+        current = next.nextResponder
+    }
+    return false
+}
+
+/// A stale default (e.g. ⌘⇧L after Open Browser is remapped) is free for
+/// web-extension commands like Bitwarden autofill while a browser web
+/// view has focus; otherwise the stale-menu-shortcut suppression would eat it.
+@MainActor
+func cmuxRouteWebExtensionCommandForStaleMenuShortcut(_ event: NSEvent, responder: NSResponder?) -> Bool {
+    guard #available(macOS 15.4, *), cmuxRespondersContainBrowserWebView(responder),
+          AppDelegate.shared?.shortcutEventBrowserPanel(event)?.performWebExtensionCommand(for: event) == true else {
+        return false
+    }
+#if DEBUG
+    cmuxDebugLog("app.sendEvent routed web-extension command before stale cmux menu shortcut")
+#endif
+    return true
+}
+
+/// Manifest commands may use Control or Option, so browser key handling
+/// dispatches them before the Command-only cmux menu-equivalent guard.
+@MainActor
+func cmuxPerformBrowserWebExtensionCommandKeyEquivalent(_ event: NSEvent) -> Bool {
+    guard #available(macOS 15.4, *),
+          let appDelegate = AppDelegate.shared,
+          appDelegate.shouldOfferBrowserWebExtensionCommand(event) else {
+        return false
+    }
+    return appDelegate.shortcutEventBrowserPanel(event)?.performWebExtensionCommand(for: event) == true
+}
+
 extension AppDelegate {
+    func matchConfiguredShortcut(event: NSEvent, shortcut: StoredShortcut) -> Bool {
+        guard !shortcut.isUnbound else { return false }
+        if let prefix = activeConfiguredShortcutChordPrefixForCurrentEvent {
+            guard let secondStroke = shortcut.secondStroke,
+                  shortcut.firstStroke == prefix else {
+                return false
+            }
+            return matchShortcutStroke(event: event, stroke: secondStroke)
+        }
+        guard !shortcut.hasChord else { return false }
+        return matchShortcutStroke(event: event, stroke: shortcut.firstStroke)
+    }
+
     /// Extension manifest commands run only after configured cmux shortcuts decline the event.
     func shouldOfferBrowserWebExtensionCommand(_ event: NSEvent) -> Bool {
         shouldOfferBrowserWebExtensionCommand(
